Part 20(0) - "Some JSR-250 Annotations".
Use the JSR-250 @PostConstruct and @PreDestroy annotations to register
init and destroy methods.

Change the type in the declaration of the context member variable in the
DrawingApp class from ApplicationContext to AbstractApplicationContext
and register a shut down hook.
Declare initialization and destruction methods in the Circle class. Mark
the initialization method with the @PostConstruct annotation and mark
the destruction method with the @PreDestroy annotation.

Signed-off-by: Bob Piper <bob.piper1953@virginmedia.com>