Part 13.0 - "Bean Definition Inheritance".
Add a parent bean for the Triangle.

Amend spring.xml by adding a parent bean for the Triangle that only
defines pointA.
Change the bean definition for the Triangle bean. Change the id to
"Triangle1" and add 'parent="parentTriangle" to its heading. Remove
pointA form the definition.
Add a second Triangle bean called "Triangle2" with the same parent as
"Triangle1". Add a property only for "pointB".
Change the DrawingApp class to get the "Triangle1" bean.

Signed-off-by: Bob Piper <bob.piper1953@virginmedia.com>