Part 17.0 - "Coding to Interfaces".
Modify the program to code using interfaces.

Add a new interface called "Shape" with a single method called "draw()".
Modify the 'Triangle' class to implement the 'Shape' interface. Modify
the draw method in 'Triangle' by adding a message that announces that a
triangle is drawn.
Add a new class called "Circle" that has two members called "centre" and
"radius". Add a draw method that announces that a circle is drawn and
outputs the centre and radius.
Modify the DrawingApp so that it instantiates two shapes called "shape1"
and "shape2", getting the beans called "shape1" and shape2" from
spring.xml and calling the draw methods of each.
Modify spring.xml by changing the name of the "Triangle" bean to
"shape1".
Modify spring.xml by adding a bean called "shape2" that refers to the
"Circle" class and defines the centre point and radius.

Signed-off-by: Bob Piper <bob.piper1953@virginmedia.com>