Part 22.0 - "Using MessageSource to Get Text From Property Files".
Use a properties file to store messages.

Create a file called "myMessages.properties" on the classpath that
contains an application greeting.
Add a message source bean to spring.xml that refers to the myMessages
file.
Add a greeting message to the DrawingApp class that fetches the
application greeting from the myMessages file.

Signed-off-by: Bob Piper <bob.piper1953@virginmedia.com>